+++ src/helpers/project.py
@@ -1,16 +1,16 @@
+import json
 import ntpath
 import os
 import subprocess
-import zipfile
-import json
 import sys
-
-from logger import Logger
-from file import File
-from searchwords import Searchwords
-import configparser
+import zipfile
 from collections import OrderedDict
 
+import configparser
+
+from helpers.file import File
+from helpers.logger import Logger
+from helpers.searchwords import Searchwords
 
 PATH = os.getcwd()
 
@@ -53,7 +53,7 @@ class Project:
                     self.db_files[full_file_name].find_matches_in_db_file()
         self.all_files.update(self.db_files)
         self.all_files.update(self.src_files)
-        from report_html import Report_html
+        from helpers.report_html import Report_html
         treeview = Report_html.Tree_builder(self, "")
         self.tree_object = treeview.return_tree_object()
