+++ src/helpers/report_html.py
@@ -1,17 +1,18 @@
-import configparser
+import base64
+import codecs
 import errno
 import hashlib
 import json
 import os
 import re
 import sqlite3
-import base64
-import codecs
 
-from html_page import Htmlpage
-from logger import Logger
-from project import Project
-from searchwords import Searchwords
+import configparser
+
+from helpers.html_page import Htmlpage
+from helpers.logger import Logger
+from helpers.project import Project
+from helpers.searchwords import Searchwords
 
 
 class Report_html(Htmlpage):
